Fixed derived type for StructuredMesh (had int4s instead of real8s for origin and width).

+++ src/mesh_header.f90
@@ -12,8 +12,8 @@ module mesh_header
      integer :: type
      integer :: n_dimension
      integer, allocatable :: dimension(:)
-     integer, allocatable :: origin(:)
-     integer, allocatable :: width(:)
+     real(8), allocatable :: origin(:)
+     real(8), allocatable :: width(:)
   end type StructuredMesh
 
 end module mesh_header
